+++ src/services/google/oauth.ts
@@ -1,7 +1,6 @@
 import type { GoogleOAuthConfig } from '@model/Config';
-import type { Identity } from '@model/Identity';
 import type { AuthorizationForIdp } from '@model/IdpAuthorization';
-import type { Email, IdpId } from '@own-types/model';
+import type { Email, Identity, IdpId } from '@notifycal/shared/types';
 import { throwError } from '@services/common/error-handling';
 import { idGenerator } from '@services/id-generator';
 import { BaseGoogle } from './base-service';